requirements.txt
streamlit>=1.30.0
google-generativeai>=0.8.0
openai>=1.0.0
pillow>=10.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
